+++ shareaza/G2Packet.cpp
@@ -1,7 +1,7 @@
 //
 // G2Packet.cpp
 //
-// Copyright (c) Shareaza Development Team, 2002-2008.
+// Copyright (c) Shareaza Development Team, 2002-2009.
 // This file is part of SHAREAZA (shareaza.sourceforge.net)
 //
 // Shareaza is free software; you can redistribute it
@@ -361,7 +361,7 @@ void CG2Packet::WriteString(LPCTSTR pszString, BOOL bNull)
 		return;
 	}
 
-	CStringA strUTF8 = UTF8Encode( pszString, _tcslen( pszString ) );
+	CStringA strUTF8 = UTF8Encode( pszString, (int)_tcslen( pszString ) );
 
 	if ( bNull )
 	{
